🐛 Fix missing `environment$` update.
The scenario of this missing updated was when restarting youwol with a different config while colab was waiting to reconnect its WS.
The source of problem was that the `environmentClient.status$` signal emitted by py-youwol was not caught because the WS was actually not yet reconnected (`confChanged$` did not emit).

Now, upon reconnection, new environment status is explicitly queried.